Completed 3 areas using well3112 branch
Greatly increased the size of the areas we already had, as well as
making them much more attractive. The areas have a lot of good space to
place items and NPCs. I also created a new cyber level. The outside area
can be accessed by entering the door on the north-west corner of the
first area, and the cyber area can be accessed by entering the door in
the north-east corner of the first area. I just added more assets and
tile classes, so I don't think there should be any issues with merging
this with everyone else's branches.